+++ app/models/Product.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class Product(id: Long, name: String, brand: String, imageUrl: String, price: Int)
+case class Product(id: Long, name: String, brand: String, imageUrl: String, price: BigDecimal)
 
 object Product {
